Add files via upload
Write efficient algorithms to solve each of the three tasks below:
1) Given an array K with N integers from 1 to N+1 such that the array has exactly one integer missing, write
a Java function that returns the missing integer.
e.g. given K = [3,5,4,1], the function should return 2
2) Given a string S of length N, write a Java function that transforms the string by reversing characters in
groups of four, and returns the transformed string.
e.g. when S = 'Lorem at' the output should be 'eroLta m'
when S = ' Tempor ip' the output should be 'meT roppi'
3) A mouse is looking over a river that's 21 feet wide. There are 20 stones placed in the river 1 foot apart,
such that there is 1 foot from the beginning of the river to the first stone, and 1 foot from the first stone
to the next stone, and so on; and 1 foot from the last stone to the end of the river. The mouse can jump
over 1 foot or 2 feet such that the first step the mouse can make is either from the beginning of the
river to the first stone, or from the beginning of the river to the second stone. If the mouse can't jump
backwards, write a Java function that returns the number of different ways the mouse can jump from
the beginning of the river to the end of the river.